+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.0.0] - 2023-07-06
+### Changed
+- Moved out type related logic to `TypeHandler`
+
+### Added
+- In `RequestArrayHandler::getEnum` BackedEnum can be used as a default now 
+
+### Fixed
+- When `RequestArrayHandler::getSingleValue` was used with a parameter holding an array, it triggered an array to (string|int/'float) conversion error. This now sets a parameter invalid error
+- When `RequestArrayHandler::getSingleValue` was simply casting anything to the desired type without really validating it. It will set an invalid parameter error now
+
+## [3.0.1] - 2023-07-03
+### Added
+- Added `getUuid` to `RequestArrayHandler`
+
+
 ## [3.0.0] - 2023-06-21
 ### Changed
 - Renamed methods in `RequestArrayHandler`, removed `FromArray` from their names
@@ -14,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - The default value used in the getter methods of `RequestArrayHandler` can be set up in the constructor 
+
 
 ## [2.0.1] - 2022-12-05
 ### Fixed